+++ examples/sketch-offline/script.js
@@ -28,7 +28,7 @@ const featureExtractor = mobileNet();
 
 const label = textInput();
 label.title = 'Instance label';
-const capture = button({ text: 'Capture this drawing' });
+const capture = button('Capture this drawing');
 capture.title = 'Capture instances to the training set';
 
 const store = dataStore('localStorage');
@@ -49,7 +49,7 @@ capture.$click
 // TRAINING
 // -----------------------------------------------------------
 
-const b = button({ text: 'Train' });
+const b = button('Train');
 
 const classifier = mlpClassifier({ layers: [64, 32], epochs: 20, dataStore: store });
 classifier.sync('sketch-classifier');
@@ -67,7 +67,7 @@ const plotTraining = trainingPlot(classifier);
 const batchMLP = batchPrediction({ name: 'mlp', dataStore: store });
 const confMat = confusionMatrix(batchMLP);
 
-const predictButton = button({ text: 'Update predictions' });
+const predictButton = button('Update predictions');
 predictButton.$click.subscribe(async () => {
   if (!classifier.ready) {
     throwError(new Error('No classifier has been trained'));