+++ examples/wizard/script.js
@@ -29,7 +29,7 @@ const featureExtractor = mobileNet();
 
 const labelInput = textInput();
 labelInput.title = 'Instance label';
-const capture = button({ text: 'Hold to record instances' });
+const capture = button('Hold to record instances');
 capture.title = 'Capture instances to the training set';
 
 const store = dataStore('localStorage');
@@ -50,7 +50,7 @@ input.$images
 // TRAINING
 // -----------------------------------------------------------
 
-const b = button({ text: 'Train' });
+const b = button('Train');
 const classifier = mlpClassifier({ layers: [64, 32], epochs: 20, dataStore: store });
 classifier.sync('wizard-classifier');
 b.$click.subscribe(() => classifier.train(trainingSet));
@@ -66,7 +66,7 @@ const plotTraining = trainingPlot(classifier);
 const batchMLP = batchPrediction({ name: 'mlp', dataStore: store });
 const confMat = confusionMatrix(batchMLP);
 
-const predictButton = button({ text: 'Update predictions' });
+const predictButton = button('Update predictions');
 predictButton.$click.subscribe(async () => {
   await batchMLP.clear();
   await batchMLP.predict(classifier, trainingSet);
@@ -76,7 +76,7 @@ predictButton.$click.subscribe(async () => {
 // REAL-TIME PREDICTION
 // -----------------------------------------------------------
 
-const tog = toggle({ text: 'toggle prediction' });
+const tog = toggle('toggle prediction');
 
 const $predictions = input.$images
   .filter(() => tog.$checked.value)
@@ -108,7 +108,7 @@ dash.settings.dataStores(store).datasets(trainingSet).models(classifier);
 // WIZARD
 // -----------------------------------------------------------
 
-const wizardButton = button({ text: 'Record Examples (class a)' });
+const wizardButton = button('Record Examples (class a)');
 const wizardText = text({ text: 'Waiting for examples...' });
 wizardButton.$pressed.subscribe((x) => {
   capture.$pressed.set(x);