+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `modelParameters` was improved to support other types
 - New widgets (components): `number` and `numberArray`
 - `textField` was updated to `textInput` and its API was updated (`$value` stream instead of `$text`)
+- `button`: minor changes to the constructor
 
 ## 0.4.4